ai-model/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
Pillow==10.0.0
yolov5==7.0.13
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
